refactor : User -> Member 로 클래스 명 변경 및 필드 추가
