--- src/main/java/codesquad/business/domain/User.java
@@ -1,43 +1,0 @@
-package codesquad.business.domain;
-
-import java.util.Objects;
-
-public class User {
-    private String userId;
-    private String username;
-    private String password;
-
-    private User(String userId, String username, String password) {
-        this.userId = userId;
-        this.username = username;
-        this.password = password;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof User user)) return false;
-        return Objects.equals(userId, user.userId) && Objects.equals(username, user.username) && Objects.equals(password, user.password);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(userId, username, password);
-    }
-
-    public String getUserId() {
-        return userId;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public static User factoryMethod(String id, String username, String password) {
-        return new User(id, username, password);
-    }
-}
